requirements.txt
# asgiref==3.8.1
# Django==5.0.6
# django-debug-panel==0.8.3
# django-debug-toolbar==4.4.2
# mysql==0.0.3
# mysqlclient==2.2.4
# pipfile==0.0.2
# sqlparse==0.5.0
# toml==0.10.2

[[source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[packages]
django = "*"
django-debug-toolbar = "*"
mysqlclient = "*"
djangorestframework = "*"
drf-nested-routers = "*"
django-filter = "*"
djoser = "*"
djangorestframework-simplejwt = "*"

[dev-packages]
autopep8 = "*"


[requires]
python_version = "3.9"
